MAGECLOUD-2901: [Spike] Prefixes for DB Tables are not Considered
